usb: chipidea: set ITC to 0 for device mode
ITC (Interrupt Threshold Control) is used to set the maximum rate at which
the host/device controller will issue interrupts. The default value is 8 (1ms)
for it. EHCI core will modify it to 1, but device mode keeps it as default
value.

In some use cases like this CR, it uses Android ADB to transfer data, ADB
only has one usb request for each direction, and maximum payload data is
only 4KB, so the speed is 4MB/s at most, it needs controller to trigger
interrupt as fast as possible to increase the speed. The USB performance
will be better if the interrupt can be triggered faster.

In this case, we set ITC default value as 0, which will trigger USB interrupt
immediately once the transfer has completed.

Signed-off-by: Peter Chen <peter.chen@freescale.com>